SelectFragment works, AddProductFragment shown

+++ Frontend/PriceTag/app/src/main/java/antunmod/projects/pricetag/RestServiceClient.java
@@ -156,7 +156,7 @@ public interface RestServiceClient {
 
     @Headers("Content-Type: application/json")
     @GET("stores/address")
-    Call<Byte> getStoreIdForAddress(@Query("storeAddress") String selectedStoreAddress);
+    Call<Short> getStoreIdForAddress(@Query("storeAddress") String selectedStoreAddress);
 
     @Headers("Content-Type: application/json")
     @GET("sizes")
