+++ Frontend/PriceTag/app/src/main/java/antunmod/projects/pricetag/transfer/AddStoreSpecificProductSpecific.java
@@ -16,10 +16,10 @@ public class AddStoreSpecificProductSpecific {
 
 	private BaseProduct baseProduct;
 	private Short productId;
-	private byte storeId;
+	private Short storeId;
 	private String storeAddress;
 
-	public AddStoreSpecificProductSpecific(BaseProduct baseProduct, Short productId, byte storeId,
+	public AddStoreSpecificProductSpecific(BaseProduct baseProduct, Short productId, Short storeId,
 			String storeAddress) {
 		super();
 		this.baseProduct = baseProduct;
@@ -28,7 +28,7 @@ public class AddStoreSpecificProductSpecific {
 		this.storeAddress = storeAddress;
 	}
 
-	public byte getStoreId() {
+	public Short getStoreId() {
 		return storeId;
 	}
